Updated label model: labelId is integer, not long. updated sql-scripts, implemented label delete-functionality, some css-mangling
git-svn-id: http://agilefant.googlecode.com/svn/trunk@3565 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/business/LabelBusiness.java
@@ -5,5 +5,6 @@ import fi.hut.soberit.agilefant.model.Story;
 
 public interface LabelBusiness extends GenericBusiness<Label> {
    
-    public long createStoryLabel(Label label, Story story);
+    public Integer createStoryLabel(Label label, Story story);
+
 }
